requirements.txt
streamlit
python-dotenv
groq
PyMuPDF
faiss-cpu
sentence-transformers
pysbd